# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BOUKHATEM TELECOM - E-commerce API for phone repair equipment
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── CATEGORIES ─────────────────────────────────────────────────────────────
...
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List active banners
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create a banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    patch:
      operationId: updateBanner
      tags: [banners]
      summary: Update a banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete a banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── SETTINGS ────────────────────────────────────────────────────────────────
...
  /admin/customers:
    get:
      operationId: listCustomers
      tags: [admin]
      summary: List all customers (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
  /admin/customers/{id}:
    get:
      operationId: getCustomer
      tags: [admin]
      summary: Get customer details (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
    patch:
      operationId: updateCustomer
      tags: [admin]
      summary: Block or update customer (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerAdminUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/sales-chart:
    get:
      operationId: getSalesChart
      tags: [admin]
      summary: Get sales chart data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesChartPoint"
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH ──────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, staff, customer]
        isBlocked:
          type: boolean
        totalOrders:
          type: number
        totalSpent:
          type: number
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string

    # ── CATEGORY ─────────────────────────────────────────────────────────────
...
    OrderItem:
      type: object
      required: [productId, name, price, quantity]
      properties:
        productId:
          type: number
        name:
          type: string
        price:
          type: number
        quantity:
          type: number
        images:
          type: array
          items:
            type: string
    PaymentMethod:
      type: string
      enum: [cash_on_delivery, bank_transfer, cib_edahabia]
    PaymentStatus:
      type: string
      enum: [pending, awaiting_confirmation, confirmed, failed]
    Order:
      type: object
      required: [id, status, subtotal, total, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        userName:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        paymentProofUrl:
          type: ["string", "null"]
        paymentNotes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        discount:
          type: number
        couponCode:
          type: ["string", "null"]
        shipping:
          type: number
        total:
          type: number
        shippingAddress:
          $ref: "#/components/schemas/Address"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Address:
      type: object
      required: [fullName, phone, wilaya, address]
      properties:
        fullName:
          type: string
        phone:
          type: string
        wilaya:
          type: string
        commune:
          type: string
        address:
          type: string

    # ── COUPON ────────────────────────────────────────────────────────────────
...
    Banner:
      type: object
      required: [id, title, isActive, sortOrder]
      properties:
        id:
          type: number
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
        buttonText:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: number
    BannerInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        subtitle:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        buttonText:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: number
    BannerUpdate:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        buttonText:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: number

    # ── SETTINGS ──────────────────────────────────────────────────────────────
...
    SalesChartPoint:
      type: object
      required: [date, sales, orders]
      properties:
        date:
          type: string
        sales:
          type: number
        orders:
          type: number
    CustomerListResponse:
      type: object
      required: [customers, total, page, totalPages]
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
    CustomerDetail:
      type: object
      required: [user, orders]
      properties:
        user:
          $ref: "#/components/schemas/User"
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
    CustomerAdminUpdate:
      type: object
      properties:
        isBlocked:
          type: boolean
        role:
          type: string
          enum: [admin, staff, customer]
...